@@ -1,14 +1,17 @@
 package ml.wonwoo.githubissues.project;
 
-import org.springframework.data.annotation.Id;
-import org.springframework.data.mongodb.core.mapping.Document;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
 
-@Document
+@Entity
 public class GithubProject {
 
   @Id
-  private String id;
+  @GeneratedValue
+  private Long id;
 
   private String orgName;
 
@@ -22,12 +25,12 @@ public class GithubProject {
     this.repoName = repoName;
   }
 
-  public String getId() {
+  public Long getId() {
     return id;
   }
 
   /* for test */
-  public void setId(String id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
