@@ -1,6 +1,6 @@
 package ml.wonwoo.githubissues.project;
 
-import org.springframework.data.mongodb.repository.MongoRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GithubProjectRepository extends MongoRepository<GithubProject, String> {
+public interface GithubProjectRepository extends JpaRepository<GithubProject, Long> {
 }
